#22 Replace handcrafted Lucene handler by "Embedded" Solr

+++ src/main/java/de/mirkosertic/desktopsearch/Facet.java
@@ -15,10 +15,10 @@ package de.mirkosertic.desktopsearch;
 public class Facet {
 
     private final String name;
-    private final int number;
+    private final long number;
     private final String link;
 
-    public Facet(String name, int number, String link) {
+    public Facet(String name, long number, String link) {
         this.name = name;
         this.number = number;
         this.link = link;
@@ -28,7 +28,7 @@ public class Facet {
         return name;
     }
 
-    public int getNumber() {
+    public long getNumber() {
         return number;
     }
